requirements.txt
python-dotenv>=0.19.1
pytest
pytest-cov
flake8